Relevant error message when static data schema mismatches
Because static data is generated by hand, sometimes an incorrect
column will be added to the static data schema. The previous code
assumed that the column names would always match the current
schema, but it is very easy to make a mistake. This change
addresses that issue by throwing runtime exceptions in such cases.