fix MigrationContext building with respect to adding indices
Previously, ADD_INDEX migrations were not getting counted when
building up the migration context. This led to the DiffGenerator
always detecting that the index needed to be added.

The solution was to update the MigrationContext.update method to
appropriately update its representation of the Column that had its
index added after it was originally created without an index.