query criteria now groupable for more complex query support
Previously, queries that used parentheses were, for the
most part, not supported. Therefore, AND chaining worked
and OR chaining worked, but it was difficult to write a
query that used both AND and OR. The solution was to force
the DBMSIntegrator to generate the WHERE string while the
Finder accumulated a list of WhereElement objects.

Additionally, the nth and nthFromLast methods were added
to enrich the limiting/offsetting API.

Additionally, there is a then() method that returns the
Resolver on the Finder so that no more dummy criteria
must be entered in order to do a distinct query or a
limit/offset query or a non-default projection.