Update DBMSIntegrator API
Previously, the string for the WHERE part of the query was built
in the Finder class of the forsuredapi. This meant it had to know
a little bit about the SQL dialect.  Now that WhereElement is
used to abstract the SQL dialect info from Finder, DBMSIntegrator
had to be updated.

In addition to the unit tests, many integration tests on a single
table were backfilled, such as complex AND/OR WHERE clauses and
LIMIT/OFFSET queries.