Remove FSSaveApi and FSDocStoreSaveApi
No longer generate FSSaveApi extensions or generate code
implementing any FSSaveApi or FSDocStoreSaveApi. It turned out
that needless code was getting generated. The lack of the need for
these interfaces was obviated by the switch to generating Setter
implementations of those interfaces instead of using reflective
proxies.

This change reduces complexity and results in one two fewer class
files in the forsuredbapi as well as one fewer class file for each
table declaration.