remove addEqualsOrChainToBuf method
The method was confusing and ultimately unnecessary. Instead of
using a specific method for adding an equals or chain, loop
through the varargs array, calling addToBuf for each value in the
array.

This comes with the additional necessity to surround all added
conditions with parentheses. Therefore, they get added to a group
first--even if there is only one value.