Feature/defaults on primitive columns if not specified (#168)
* assume java default for primitives to avoid NPE

There would have been an NPE when taking null values
from the database and assigning to Java primitives. This
change prevents that from happening by inserting the 
Java defaults as the database defaults if not otherwise
defined. Thus, from this point on, if a column should be
allowed to be null, then use a primitive wrapper.