remove legacy tests that will not matter after redesign
Previously, the output of the differencing operation was a
MigrationSet which had an ordered list of migrations. The new
design has no such ordered list of migrations, but MigrationSet
and Migration still both exists because existing projects will
have serialized them and will need them at runtime.

The eventual goal is to serialize the difference between the
previous and next database schema. Therefore, any tests where
Migration is an input or an expected output are no longer valid.